Add citation collapse delimiter for year
Update the citation layout so that the delimiter used for a single citation with one author but multiple years is a ';'. The use of the semicolon is consistent with the layout produced for a single citation with multiple author year pairs. For example the current version would have the citation as (Schrijver 2015, 2018) whereas the edited version is (Schrijver 2015; 2018).